Feat: 방 입장
특정한 회원이 특정한 방에 입장하는 부분 완료, 승인 부분 고민 필요
현재 게임 시작 부분 구현 중(마니또-마니띠 랜덤 매칭)

+++ backend/src/main/java/com/sww/ddorangddorang/domain/room/api/RoomApi.java
@@ -18,7 +18,7 @@ public class RoomApi {
     private final RoomService roomService;
 
     @PostMapping("/")
-    public Integer createRoom(@RequestHeader Integer userId, @RequestBody RoomInfoReq roomInfoReq) {
+    public Integer createRoom(@RequestHeader Long userId, @RequestBody RoomInfoReq roomInfoReq) {
         log.info("RoomApi_createRoom start");
         Integer accessCode = roomService.createRoom(userId, roomInfoReq);
         log.info("RoomApi_createRoom end: " + accessCode);
